fix: unit test based on refactored code
